Fixed protottype.
Rev: src/interpret.h:1.62

+++ src/interpret.h
@@ -5,7 +5,7 @@
 \*/
 
 /*
- * $Id: interpret.h,v 1.61 2000/08/08 19:38:39 grubba Exp $
+ * $Id: interpret.h,v 1.62 2000/08/09 10:27:41 grubba Exp $
  */
 #ifndef INTERPRET_H
 #define INTERPRET_H
@@ -213,7 +213,7 @@ do{ \
 #include "block_alloc_h.h"
 /* Prototypes begin here */
 void push_sp_mark(void);
-int pop_sp_mark(void);
+ptrdiff_t pop_sp_mark(void);
 void init_interpreter(void);
 void lvalue_to_svalue_no_free(struct svalue *to,struct svalue *lval);
 void assign_lvalue(struct svalue *lval,struct svalue *from);
